chore: make benchmarks not run (the main) when running go test ./... on root repo
